more refactoring

+++ engine/src/controller.rs
@@ -1,5 +1,5 @@
 use crate::state::player::Player;
 
 pub trait Controller {
-    fn should_stop(&self, in_search: bool, side: Player, nodes: u128, ply_or_depth: u8) -> bool;
+    fn should_stop(&self, in_search: bool, side: Player, nodes: u64, ply_or_depth: u8) -> bool;
 }
